+++ Firebase/Storage/FIRStorageObservableTask.m
@@ -102,7 +102,7 @@
       break;
 
     case FIRStorageTaskStatusUnknown:
-    // Fall through to exception case if an unknown status is passed
+      // Fall through to exception case if an unknown status is passed
 
     default:
       [NSException raise:NSInternalInconsistencyException