+++ Firestore/Source/Local/FSTLevelDB.mm
@@ -19,16 +19,23 @@
 #include <leveldb/db.h>
 
 #import "FIRFirestoreErrors.h"
-#import "Firestore/Source/Core/FSTDatabaseInfo.h"
+#import "Firestore/Source/Local/FSTLevelDBMigrations.h"
 #import "Firestore/Source/Local/FSTLevelDBMutationQueue.h"
 #import "Firestore/Source/Local/FSTLevelDBQueryCache.h"
 #import "Firestore/Source/Local/FSTLevelDBRemoteDocumentCache.h"
 #import "Firestore/Source/Local/FSTWriteGroup.h"
 #import "Firestore/Source/Local/FSTWriteGroupTracker.h"
-#import "Firestore/Source/Model/FSTDatabaseID.h"
 #import "Firestore/Source/Remote/FSTSerializerBeta.h"
 #import "Firestore/Source/Util/FSTAssert.h"
 #import "Firestore/Source/Util/FSTLogger.h"
+
+#include "Firestore/core/src/firebase/firestore/core/database_info.h"
+#include "Firestore/core/src/firebase/firestore/model/database_id.h"
+#include "Firestore/core/src/firebase/firestore/util/string_apple.h"
+
+namespace util = firebase::firestore::util;
+using firebase::firestore::core::DatabaseInfo;
+using firebase::firestore::model::DatabaseId;
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -36,6 +43,7 @@ static NSString *const kReservedPathComponent = @"firestore";
 
 using leveldb::DB;
 using leveldb::Options;
+using leveldb::ReadOptions;
 using leveldb::Status;
 using leveldb::WriteOptions;
 
@@ -49,6 +57,15 @@ using leveldb::WriteOptions;
 @end
 
 @implementation FSTLevelDB
+
+/**
+ * For now this is paranoid, but perhaps disable that in production builds.
+ */
++ (const ReadOptions)standardReadOptions {
+  ReadOptions options;
+  options.verify_checksums = true;
+  return options;
+}
 
 - (instancetype)initWithDirectory:(NSString *)directory
                        serializer:(FSTLocalSerializer *)serializer {
@@ -72,13 +89,13 @@ using leveldb::WriteOptions;
 
 #else
 #error "local storage on tvOS"
-// TODO(mcg): Writing to NSDocumentsDirectory on tvOS will fail; we need to write to Caches
-// https://developer.apple.com/library/content/documentation/General/Conceptual/AppleTV_PG/
+  // TODO(mcg): Writing to NSDocumentsDirectory on tvOS will fail; we need to write to Caches
+  // https://developer.apple.com/library/content/documentation/General/Conceptual/AppleTV_PG/
 
 #endif
 }
 
-+ (NSString *)storageDirectoryForDatabaseInfo:(FSTDatabaseInfo *)databaseInfo
++ (NSString *)storageDirectoryForDatabaseInfo:(const DatabaseInfo &)databaseInfo
                            documentsDirectory:(NSString *)documentsDirectory {
   // Use two different path formats:
   //
@@ -88,11 +105,14 @@ using leveldb::WriteOptions;
   // projectIDs are DNS-compatible names and cannot contain dots so there's
   // no danger of collisions.
   NSString *directory = documentsDirectory;
-  directory = [directory stringByAppendingPathComponent:databaseInfo.persistenceKey];
+  directory = [directory
+      stringByAppendingPathComponent:util::WrapNSStringNoCopy(databaseInfo.persistence_key())];
 
-  NSString *segment = databaseInfo.databaseID.projectID;
-  if (![databaseInfo.databaseID isDefaultDatabase]) {
-    segment = [NSString stringWithFormat:@"%@.%@", segment, databaseInfo.databaseID.databaseID];
+  NSString *segment = util::WrapNSStringNoCopy(databaseInfo.database_id().project_id());
+  if (!databaseInfo.database_id().IsDefaultDatabase()) {
+    segment = [NSString
+        stringWithFormat:@"%@.%@", segment,
+                         util::WrapNSStringNoCopy(databaseInfo.database_id().database_id())];
   }
   directory = [directory stringByAppendingPathComponent:segment];
 
@@ -115,8 +135,8 @@ using leveldb::WriteOptions;
   if (!database) {
     return NO;
   }
-
   _ptr.reset(database);
+  [FSTLevelDBMigrations runMigrationsOnDB:_ptr];
   return YES;
 }