+++ cmake/xcodebuild.cmake
@@ -33,7 +33,7 @@ function(xcodebuild framework)
   set(options "")
   set(single_value SCHEME WORKSPACE)
   set(multi_value DEPENDS)
-  cmake_parse_arguments(xcb "${options}" "${single_value}" "${multi_value}")
+  cmake_parse_arguments(xcb "${options}" "${single_value}" "${multi_value}" ${ARGN})
 
   if(NOT xcb_WORKSPACE)
     set(xcb_WORKSPACE ${PROJECT_SOURCE_DIR}/Example/Firebase.xcworkspace)
@@ -45,8 +45,6 @@ function(xcodebuild framework)
   set(platform macOS)
   set(destination "platform=macOS,arch=x86_64")
   set(scheme "${framework}-${platform}")
-
-  set(binary_dir ${PROJECT_BINARY_DIR}/${scheme})
 
   # CMake has a variety of release types, but Xcode has just one by default.
   if(CMAKE_BUILD_TYPE STREQUAL Debug)
@@ -65,11 +63,11 @@ function(xcodebuild framework)
     ${framework}
     DEPENDS ${xcb_DEPENDS}
 
-    PREFIX ${binary_dir}
+    PREFIX ${PROJECT_BINARY_DIR}/external/${framework}
 
     # The source directory doesn't actually matter
     SOURCE_DIR ${PROJECT_SOURCE_DIR}
-    BINARY_DIR ${binary_dir}
+    BINARY_DIR ${PROJECT_BINARY_DIR}/Frameworks
 
     CONFIGURE_COMMAND ""
 
@@ -79,7 +77,7 @@ function(xcodebuild framework)
         -scheme ${scheme}
         -configuration ${configuration}
         -destination ${destination}
-        CONFIGURATION_BUILD_DIR=${FIREBASE_INSTALL_DIR}/Frameworks
+        CONFIGURATION_BUILD_DIR=<BINARY_DIR>
         build
         ${pipe_xcpretty}
     BUILD_ALWAYS ${BUILD_PODS}
@@ -87,4 +85,5 @@ function(xcodebuild framework)
     INSTALL_COMMAND ""
     TEST_COMMAND ""
   )
+
 endfunction()