+++ Firebase/Auth/CHANGELOG.md
@@ -1,3 +1,8 @@
+# v4.4.1
+- Fixes bug where the FIRAuthResult object returned following a Phone Number authentication
+  always contained a nil FIRAdditionalUserInfo object. Now the FIRAdditionalUserInfo object is
+  never nil and its newUser field is populated correctly.
+
 # v4.4.0
 - Adds new APIs which return an AuthDataResult object after successfully creating an
   Email/Password user, signing in anonymously, signing in with Email/Password and signing